workable empty store created, test code removed from App
